Move OData handler schema composition to separate file
